chore: Created a shared test setup (glint)
